Split tool artifact surface
